+++ NEWS.md
@@ -8,6 +8,7 @@
 *   Improved plotting for "ar" and "var" families.
 *   Add test coverage for `cp_only = TRUE` and `fastcpd_ts`.
 *   Increase test coverage.
+*   Provide user selection when `ggplot2` is not installed.
 
 # fastcpd 0.8.2